Update README.md
This project is designed to securely anonymize and format electronic health records (EHR) for interaction with OpenAI's API, specifically for use with the ChatGPT model. The primary aim is to facilitate a safe and efficient way for healthcare providers and researchers to gain insights from EHR data using advanced AI-driven queries and analysis.